fixing error

+++ 0x15-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-int file;
+ssize_t file;
 ssize_t r;
 ssize_t w;
 char *buffer;
